+++ R/zzz.R
@@ -17,7 +17,7 @@
       #While developing:
         packageStartupMessage(
          "#######################################################\n",
-                              "This Version 2026 05 16 - 11.47AM\n" )
+                              "This Version 2026 05 19 - 04.57AM\n" )